requirements.txt
requests
feedparser
pyyaml
jinja2
rapidfuzz
python-dateutil
openreview-py
yake
